Replace cardano-ping executable with cardano-ping library
