+++ cardano-ping/CHANGELOG.md
@@ -1,0 +1,5 @@
+# Revision history for cardano-ping
+
+## 0.1.0.0 -- 2022-12-14
+
+* This code was originally from the cardano-ping executable component of the `network-mux` package.
