+++ ouroboros-consensus-shelley-test/ouroboros-consensus-shelley-test.cabal
@@ -109,6 +109,7 @@ test-suite test
                      , ouroboros-consensus-shelley
                      , ouroboros-consensus-shelley-test
 
+  build-tool-depends:  alex:alex
   default-language:    Haskell2010
   ghc-options:         -Wall
                        -Wcompat
